Fix switch markup tests
